scripts/requirements.txt
requests==2.31.0
python-frontmatter==1.0.0
pyyaml==6.0.1
